+++ src/core/mac/mac.cpp
@@ -38,7 +38,7 @@
 #include "crypto/aes_ccm.hpp"
 #include "crypto/sha256.hpp"
 #include "instance/instance.hpp"
-#include "utils/static-counter.hpp"
+#include "utils/static_counter.hpp"
 
 namespace ot {
 namespace Mac {
