+++ src/core/thread/mesh_forwarder.cpp
@@ -34,7 +34,7 @@
 #include "mesh_forwarder.hpp"
 
 #include "instance/instance.hpp"
-#include "utils/static-counter.hpp"
+#include "utils/static_counter.hpp"
 
 namespace ot {
