+++ src/core/thread/mle.cpp
@@ -34,7 +34,7 @@
 #include "mle.hpp"
 
 #include "instance/instance.hpp"
-#include "utils/static-counter.hpp"
+#include "utils/static_counter.hpp"
 
 namespace ot {
 namespace Mle {
