+++ src/core/utils/static_counter.hpp
